Address review: dedupe the _IsInstance test matcher
- Move the _IsInstance argument-matcher into a shared tests/_matchers.py
  and import it from test_reorient_exec.py and test_worktree_index.py
  instead of duplicating the class in each; add the tests dir to sys.path
  in conftest so the sibling module resolves.
